Rename `server_tick` module into `tick`
